requirements.txt
discord.py==1.7.3
wheel==0.37.0
psycopg2==2.9.1